+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added the log of the estimated odometry from `yarp::dev::Nav2D::ILocalization2D`.
+- Deprecated the `logAllQuantities` option in favour of `logControlBoardQuantities`.
+
 ## [0.2.0] - 2021-05-19
 
 - Added the possibility to specify the time of a ``Record`` with ``push_back``.